Move `trim_and_truncate` to the source file
